backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
pydantic[email]

SQLAlchemy
psycopg[binary]

passlib[bcrypt]
bcrypt==4.1.2
python-jose

PyPDF2
python-docx

openai
